Fix up scope resolver more, add a test
---
